Move the Ssurgeon reindex operations from AddDep to SsurgeonUtils
